requirements.txt
# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
pymongo==4.6.1
motor==3.3.2

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
uvicorn==0.24.0
fastapi==0.104.1
